Changed: Moved TreeFolderGeneratorCompositeItemModelAdapter per style request in review.
