+++ tests/NexusMods.UI.Tests/Helpers/TreeDataGrid/FolderGenerator/TreeFolderGeneratorCompositeItemModelAdapterTests.cs
@@ -1,6 +1,5 @@
 using DynamicData;
 using FluentAssertions;
-using NexusMods.App.UI.Helpers.TreeDataGrid.New.FolderGenerator.Helpers;
 using NexusMods.App.UI.Helpers.TreeDataGrid.New.FolderGenerator;
 using NexusMods.App.UI.Controls;
 using NexusMods.MnemonicDB.Abstractions;
